fix(M-04): encode lovelace and fee as BigInt in CBOR
Convert netAmount and feeAmount to BigInt before passing them to cbor2 in
buildPayload. JS Number only safely represents integers up to 2^53; on-chain
lovelace amounts can in theory exceed that bound. Using BigInt forces cbor2
to emit CBOR unsigned integers (major type 0) without precision loss.

+++ src/utils/staking.utils.ts
@@ -305,9 +305,10 @@ export const buildPayload = (
   const { toAddress, txInputs, ttl, certificates, withdrawals, votingProcedures, network } =
     options;
 
-  // Ensure lovelace amounts are CBOR unsigned integers, not floats.
-  const netAmount = Math.round(options.netAmount);
-  const feeAmount = Math.round(options.feeAmount);
+  // Encode lovelace amounts as BigInt so cbor2 emits CBOR unsigned ints
+  // (major type 0) without risking JS Number (53-bit) precision loss.
+  const netAmount = BigInt(Math.round(options.netAmount));
+  const feeAmount = BigInt(Math.round(options.feeAmount));
 
   // Build inputs array
   const inputsArr = txInputs.map((input) => {
